docs/implementation/python/pyproject.toml
[tool.poetry]
name = "pyfate"
version = "0.2.4"
description = ""
authors = ["vduseev <vagiz@duseev.com>"]
license = "MIT"

[tool.poetry.scripts]
fate = "pyfate.main:console"
fate-webservice = "pyfate.main:webservice"

[tool.poetry.dependencies]
python = "^3.7"
docker = "^4.1.0"
loguru = "^0.4.1"
colorama = "^0.4.3"
python-configuration = "^0.7.1"
cleo = "^0.8.1"
pygments = "^2.7.0"
durable_rules = "^2.0.28"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^5.3.5"
behave = "^1.2.6"
flake8 = "^3.8.3"
ipython = "^7.18.1"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
